Refactored utils
