+++ src/utils/index.js
@@ -1,140 +1,13 @@
-import moment from "moment";
+export {getRandomIntegerNumber} from "./random-integer";
+export {getRandomArrayItem} from "./random-array-item";
+export {getRandomDate} from "./random-date";
+export {getRandomBoolean} from "./random-boolean";
+export {formatDate} from "./format-date";
+export {formatTime} from "./format-time";
+export {isDateExpired} from "./date-expired";
+export {isDateToday} from "./current-date";
+export {hasSomeBoolean} from "./has-some-boolean";
+export {createElement} from "./create-element";
+export {render, RenderPosition} from "./render";
+export {replace} from "./replace";
 
-const DATE_MAX_RANGE = 7;
-
-/**
- * Returns random array element
- * @param {Array} array - array to return element from
- * @return {*} - random element of the array
- */
-const getRandomArrayItem = (array) => array[getRandomIntegerNumber(0, array.length)];
-
-/**
- * Returns random integer between min (inclusive) and max (exclusive) values
- * @param {Number} min - min value
- * @param {Number} max - max value
- * @return {Number} - random integer number
- */
-const getRandomIntegerNumber = (min = 0, max = 10) => min + Math.floor((max - min) * Math.random());
-
-/**
- * Returns random date that lies within the given range from the current date
- * @param {Number} range - range from the current date
- * @return {Date} - random date
- */
-const getRandomDate = (range = DATE_MAX_RANGE) => {
-  const targetDate = new Date();
-  const sign = getRandomBoolean() ? 1 : -1;
-  const diffValue = sign * getRandomIntegerNumber(0, range);
-
-  targetDate.setDate(targetDate.getDate() + diffValue);
-
-  return targetDate;
-};
-
-/**
- * Returns random boolean value (true or false)
- * @return {Boolean} - random true or false
- */
-const getRandomBoolean = () => Math.random() > 0.5;
-
-/**
- * Formates date to string
- * @param {Date} date - date object
- * @return {String} - formated date
- */
-const formatDate = (date) => moment(date).format(`DD MMMM`);
-
-/**
- * Formates time to string
- * @param {Date} date - date object
- * @return {String} - formated time
- */
-const formatTime = (date) => moment(date).format(`hh:mm A`);
-
-/**
- * Checks if date is expired
- * @param {Date} date - date to be checked
- * @return {Boolean} - is date expired
- */
-const isDateExpired = (date) => date instanceof Date && date < Date.now();
-
-/**
- * Checks if date is due today
- * @param {Date} date - date to be checked
- * @return {Boolean} - is date due today
- */
-const isDateToday = (date) => date instanceof Date && date.toDateString() === new Date().toDateString();
-
-/**
- * Returns true if object has some Boolean property
- * @param {Object} obj - some object
- * @return {Boolean} - true if object has some Boolean property
- */
-const hasSomeBoolean = (obj) => Object.values(obj).some(Boolean);
-
-/**
- * Creates HTML element from given template
- * @param {String} template - template to create HTML element from
- * @return {HTMLElement} - created element
- */
-const createElement = (template) => {
-  const newElement = document.createElement(`div`);
-  newElement.innerHTML = template;
-
-  return newElement.firstChild;
-};
-
-const RenderPosition = {
-  AFTER_BEGIN: `afterbegin`,
-  BEFORE_END: `beforeend`
-};
-
-/**
- * Renders given component to the DOM by adding it to the parent container
- * at the specified position
- * @param {HTMLElement} container - parent HTML element
- * @param {Component} component - component to be added to the container
- * @param {String} place - insert position. Default value = "beforeend"
- */
-const render = (container, component, place = RenderPosition.BEFORE_END) => {
-  switch (place) {
-    case RenderPosition.AFTER_BEGIN:
-      container.prepend(component.getElement());
-      break;
-    case RenderPosition.BEFORE_END:
-      container.append(component.getElement());
-      break;
-  }
-};
-
-/**
- * Replaces two given components
- * @param {Component} newComponent - repacing component
- * @param {Component} oldComponent  - replaced component
- */
-export const replace = (newComponent, oldComponent) => {
-  const parentElement = oldComponent.getElement().parentElement;
-  const newElement = newComponent.getElement();
-  const oldElement = oldComponent.getElement();
-
-  const isExistElements = !!(parentElement && newElement && oldElement);
-
-  if (isExistElements && parentElement.contains(oldElement)) {
-    parentElement.replaceChild(newElement, oldElement);
-  }
-};
-
-export {
-  getRandomArrayItem,
-  getRandomDate,
-  getRandomBoolean,
-  formatDate,
-  formatTime,
-  isDateExpired,
-  isDateToday,
-  hasSomeBoolean,
-  createElement,
-  render,
-  RenderPosition
-};
