+++ src/utils/random-array-item.js
@@ -1,0 +1,8 @@
+import {getRandomIntegerNumber} from "./index";
+
+/**
+ * Returns random array element
+ * @param {Array} array - array to return element from
+ * @return {*} - random element of the array
+ */
+export const getRandomArrayItem = (array) => array[getRandomIntegerNumber(0, array.length)];
