+++ src/utils/random-boolean.js
@@ -1,0 +1,5 @@
+/**
+ * Returns random boolean value (true or false)
+ * @return {Boolean} - random true or false
+ */
+export const getRandomBoolean = () => Math.random() > 0.5;
