+++ src/utils/replace.js
@@ -1,0 +1,17 @@
+
+/**
+ * Replaces two given components
+ * @param {Component} newComponent - repacing component
+ * @param {Component} oldComponent  - replaced component
+ */
+export const replace = (newComponent, oldComponent) => {
+  const parentElement = oldComponent.getElement().parentElement;
+  const newElement = newComponent.getElement();
+  const oldElement = oldComponent.getElement();
+
+  const isExistElements = !!(parentElement && newElement && oldElement);
+
+  if (isExistElements && parentElement.contains(oldElement)) {
+    parentElement.replaceChild(newElement, oldElement);
+  }
+};
